// src/api/models/user-settings.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface UserSettings
 */
export interface UserSettings {
    /**
     * 
     * @type {boolean}
     * @memberof UserSettings
     */
    'email_notifications'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UserSettings
     */
    'in_app_notifications'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UserSettings
     */
    'dark_mode'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof UserSettings
     */
    'language'?: UserSettingsLanguageEnum;
}

export const UserSettingsLanguageEnum = {
    En: 'en',
    Fr: 'fr',
    Es: 'es'
} as const;

export type UserSettingsLanguageEnum = typeof UserSettingsLanguageEnum[keyof typeof UserSettingsLanguageEnum];